requirements.txt
pillow
pyheif 